+++ api/RELEASE_NOTES.md
@@ -5,6 +5,7 @@
 * Adds `DxApi.addTags` method
 * Fixes `DxFindDataObjects` when used with `tags` constraint
 * Handles record results in `DxFindDataObjects`
+* Adds `systemRequirements` to `DxWorkflowStageDesc`
 
 ## 0.12.0 (2021-11-18)